requirements.txt
aiogram==3.12.0
python-dotenv==1.0.1
APScheduler==3.10.4
aiosqlite==0.20.0
pytz==2024.2
